# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "sftoolApi",
    "description": "The API deliver information for the Single Project Business locations",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/spb/v1/secure/:id": {
      "get": {
        "tags": [
          "Single Project Business"
        ],
        "summary": "",
        "operationId": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "ID from the Object"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/SingleProjectBusiness"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "singleProjectBusiness": [
              "readFullSingleProjectBusiness",
              "write:singleProjectBusiness",
              "update:singleProjectBusiness",
              "delete:singleProjectBusiness"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Single Project Business"
        ],
        "summary": "Update Single Project Business",
        "operationId": "updateSingleProjectBusiness",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "ID from the Object that need to be updated"
          },
          {
            "name": "body",
            "description": "Single Project Business Object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SingleProjectBusiness"
            },
            "in": "body"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/SingleProjectBusiness"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable"
          }
        },
        "security": [
          {
            "singleProjectBusiness": [
              "readFullSingleProjectBusiness",
              "write:singleProjectBusiness",
              "update:singleProjectBusiness",
              "delete:singleProjectBusiness"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Single Project Business"
        ],
        "summary": "Delete a Single Project Business location",
        "operationId": "deleteSingleProjectBusiness",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "singleProjectBusiness": [
              "readFullSingleProjectBusiness",
              "write:singleProjectBusiness",
              "update:singleProjectBusiness",
              "delete:singleProjectBusiness"
            ]
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "SingleProjectBusiness": {
      "description": "For Auth users is the contact information enabled",
      "required": [
        "address",
        "island",
        "processingStatus"
      ],
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "readOnly": true,
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        },
        "address": {
          "required": [
            "street",
            "streetNumber",
            "zip",
            "city"
          ],
          "type": "object",
          "properties": {
            "street": {
              "type": "string"
            },
            "streetNumber": {
              "type": "string"
            },
            "zip": {
              "type": "number"
            },
            "city": {
              "type": "string"
            }
          }
        },
        "socketState": {
          "type": "string",
          "enum": [
            "OTO Ready",
            "Fiber Ready",
            "no state",
            "Unknown"
          ],
          "default": "Unknown"
        },
        "island": {
          "type": "string",
          "enum": [
            "Central",
            "West",
            "East",
            "South"
          ]
        },
        "processingStatus": {
          "type": "string",
          "enum": [
            "Construction",
            "Approved",
            "Supported"
          ]
        },
        "contact": {
          "type": "object",
          "properties": {
            "supervisor": {
              "type": "string"
            },
            "representative": {
              "type": "string"
            },
            "rla": {
              "type": "string"
            }
          }
        },
        "connectKit": {
          "type": "string",
          "enum": [
            "Yes",
            "No",
            "Unknown"
          ],
          "default": "Unknown"
        },
        "enterDate": {
          "type": "string",
          "format": "date"
        },
        "stagesState": {
          "type": "boolean",
          "default": "false"
        },
        "socketHyperlink": {
          "type": "string",
          "description": "The URL is encoded"
        },
        "comment": {
          "type": "string"
        },
        "create": {
          "type": "string",
          "format": "date",
          "readOnly": true
        },
        "update": {
          "type": "string",
          "format": "date",
          "readOnly": true
        }
      }
    }
  }
}